[REFACTOR][OrcJIT] Extract Slab from ArenaJITLinkMemoryManager
Stage A of the slab-pool refactor in refactor_plan.md. Pure internal
restructuring; zero behavior change.

- Move every per-arena concern — mmap reservation, dual-pool bump
  allocator, commit bitmap, free list with coalescing, page-protection
  primitives, InFlightAlloc logic — into a new `Slab` class in
  `orcjit_slab.{h,cc}`. Slab is the unit-of-VA-reservation; future
  stages will introduce a pool of them per session.
- `ArenaJITLinkMemoryManager` becomes a thin ~30-line wrapper that
  owns one `unique_ptr<Slab>` and delegates `allocate` / `deallocate`.
  The capacity-halving retry loop (negotiation between caller request
  and kernel RLIMIT_AS) stays here — it is not a Slab concern.
- `FinalizedAllocInfo` gains a `Slab* owner` field, stamped at finalize
  time. Redundant today (one Slab per session) but makes Stage B's
  pool-manager routing O(1) without address comparisons.
- Rename the 2 MB commit-granularity constant `kSlabSize` →
  `Slab::kCommitGranularity`. This frees the "Slab" name for the new
  class and clarifies that the constant describes page-commit / THP
  granularity, not the Slab-as-pool-unit.
- Rename the user-facing parameter `arena_size` / `arena_size_bytes`
  → `slab_size` / `slab_size_bytes` across the C++ constructors,
  Python kwarg, registered-global-func lambda, and tests. Stage A has
  one Slab per session so the semantics are identical; Stage B
  preserves the name while adding multi-slab pool behavior. This is
  an API break: `ExecutionSession(arena_size=...)` must become
  `ExecutionSession(slab_size=...)`. Doing the rename now avoids a
  second break once Stage B lands.

Verification: `pytest addons/tvm_ffi_orcjit/tests` — 64 passed, 3
skipped (identical to pre-refactor). `nm` confirms `Slab::allocate`,
`Slab::deallocateOne`, etc. in the built .so.
