+++ addons/tvm_ffi_orcjit/CMakeLists.txt
@@ -40,7 +40,8 @@ find_package(tvm_ffi CONFIG REQUIRED)
 add_library(
   tvm_ffi_orcjit SHARED
   src/ffi/orcjit_session.cc src/ffi/orcjit_dylib.cc src/ffi/orcjit_memory_manager.cc
-  src/ffi/llvm_patches/gotpcrelx_fix.cc src/ffi/llvm_patches/init_fini_plugin.cc
+  src/ffi/orcjit_slab.cc src/ffi/llvm_patches/gotpcrelx_fix.cc
+  src/ffi/llvm_patches/init_fini_plugin.cc
 )
 set_target_properties(
   tvm_ffi_orcjit PROPERTIES CXX_VISIBILITY_PRESET hidden VISIBILITY_INLINES_HIDDEN ON
